bot/src/requirements.txt
asyncio
aiogram
google-cloud-storage
requests